requirements.txt
streamlit>=1.19.0
openai>=0.27.0
PyPDF2>=3.0.1
faiss-cpu>=1.7.3
sentence-transformers>=2.2.2
numpy>=1.23.5
torch>=1.13.1  # Required for sentence-transformers if using an embedding model